fixed label font size: from int to float

+++ cocos/2d/CCLabelTTF.cpp
@@ -120,7 +120,7 @@ const std::string& LabelTTF::getString() const
 
 std::string LabelTTF::getDescription() const
 {
-    return StringUtils::format("<LabelTTF | FontName = %s, FontSize = %d, Label = '%s'>", _renderLabel->getFontName().c_str(), _renderLabel->getFontSize(), _renderLabel->getString().c_str());
+    return StringUtils::format("<LabelTTF | FontName = %s, FontSize = %f, Label = '%s'>", _renderLabel->getFontName().c_str(), _renderLabel->getFontSize(), _renderLabel->getString().c_str());
 }
 
 TextHAlignment LabelTTF::getHorizontalAlignment() const
@@ -156,7 +156,7 @@ void LabelTTF::setDimensions(const Size &dim)
     _contentDirty = true;
 }
 
-int LabelTTF::getFontSize() const
+float LabelTTF::getFontSize() const
 {
     return _renderLabel->getFontSize();
 }
